# video: Add ewa_lanczos upscaler (aka jinc)