This is the polar (elliptic weighted average) version of lanczos.
This introduces a general new form of polar/2D filters.

This commit probably still needs a bit of work before it can be merged, but I've created a pull request to keep track of feedback and make it visible.