@haasn 2015-01-09
Seems like this version is actually slower than the previous hacked-together version. The culprit seems to be the usage of sampler3D instead of sampler2D.

Based on this, I tried going even smaller and using a sampler1D that just caches the jinc function (with the weights matrix computed in realtime on the GPU). Turned out to be even faster.

Clearly, this still needs some work.