@haasn 2015-01-13
I've pushed a rework of this commit that uses a much faster 1D LUT approach. With this version, I can actually watch 60 Hz 4K videos in realtime even with ewa_lanczos4, which is a quite impressive improvement over the previous, extremely slow version. Speedup is almost 100% on my machine.