@wm4 2015-01-13
Does using a 1D texture vs. 2D texture make a difference?