@haasn 2015-01-13
@wm4 Depends on what you mean:

1. Using a 2D texture for the LUT (with x and y each mapped to a coordinate) is slower than using a 1D texture (with sqrt(xx+yy) mapped to a coordinate).

The performance numbers I get for this on my machine is something like: 12 ms (1D), 15 ms (2D) and 19ms (3D).

2. Using a 2D texture of size Nx1 to simulate a 1D texture seems to be even slower, I get about 17ms avg GPU time. (Even slower if I use 1xN, for some reason)