@wm4 2015-01-13
I've recently replaced the 1D LUTs for the separable scalers with a 1xN LUT (1 pixel width), and found no measurable speed difference. This was on a GeForce 8500 GT with nvidia drivers.