@haasn 2015-01-13
Which scalers does that refer to? None of the separable scalers in the current code use 1D LUTs, at least as far as I can see.

(I've rebased and squashed the commit, for easier access to the new code)