@wm4 2015-01-13
>Which scalers does that refer to? None of the separable scalers in the current code use 1D LUTs, at least as far as I can see.

The convolution filters. E.g. ``weights2()`` was implemented using a 1D texture. This was changed in 4a95be014b8875567e8e800a6075d876ba824081. Maybe you could test whether the commit reduces performance for you?