@haasn 2015-01-13
I ran each test a few times to eliminate margin of error as much as possible. The numbers I'm reporting here are the ones from the first program listed by qapitrace, which, if I'm not mistaken, corresponds to the second pass only for the separable filters (it has 8 avg million pixel drawn, and the highest time taken).

4cae83d, lanczos3: 1.795ms, 1.799ms, 1.792ms, 1.806ms, 1.792ms
4a95be0, lanczos3: 1.799ms, 1.806ms, 1.799ms, 1.799ms, 1.799ms

4cae83d, lanczos2: 1.138ms, 1.313ms, 1.120ms, 1.482ms, 1.120ms
4a95be0, lanczos2: 1.132ms, 1.305ms, 1.156ms, 1.163ms, 1.139ms

I'm not sure what causes the high amount of fluctuations in particular in the lanczos2 case, but it doesn't seem like it significantly impacts performance here. To eliminate any similar margin of error from my other test, here is the exact patch used and the numbers I get out of them.

unpatched, ewa_lanczos4: 11.988ms, 11.949ms, 11.961ms, 11.918ms
unpatched, ewa_lanczos2: 3.793ms, 3.794ms, 3.819ms, 3.839ms, 3.808ms

This is the first version of the patch, for a 1xN texture: https://bpaste.net/show/a80e0fa8e8e5

patched 1xN, ewa_lanczos4: 20.485ms, 20.463ms, 20.502ms, 20.462ms
patched 1xN, ewa_lanczos2: 5.607ms, 5.608ms, 5.619ms, 5.611ms, 5.607ms

This is the second version of the patch, for an Nx1 texture: https://bpaste.net/show/2b0905cfa371

patched Nx1, ewa_lanczos4: 17.303ms, 17.308ms, 17.304ms, 17.310ms
patched Nx1, ewa_lanczos2: 5.181ms, 5.154ms, 5.153ms, 5.158ms, 5.154ms

There is a clearly noticeable speed loss here. Maybe it's related to the number of texture channels? In each case, I am still using monochrome textures (in GL_RED format).