@wm4 2015-01-13
Huh, that's really weird. I can't make any sense out of this. Just for the record, what's your gpu/driver?

We could probably just leave it as is, but still makes me feel uneasy, and it makes the code more complex.