@haasn 2015-01-13
My GPU is AMD Radeon HD 7950 (900 MHz) and the exact version information is:

```
Package: x11-drivers/ati-drivers-14.9-r1
Catalyst version: 12.3
Driver packaging version: 8.951-120308a-135854C-ATI
2D Driver version: 14.30.4
Catalyst control center version: 2.21
RandR version: 1.3
```

(Not really sure how to interpret those version numbers)